requirements.txt
Django==1.8.3
django-suit==0.2.15
numpy==1.10.1
pHash==0.1
Pillow==3.0.0
wheel==0.24.0
